requirements.txt
Django>=5.0.4,<5.1.0
djangorestframework>=3.14.0,<3.15.0
drf-yasg>=1.21.7,<1.22.0
psycopg2-binary>=2.9.9,<2.10.0
gunicorn>=21.2.0,<21.3.0
python-decouple>=3.8,<3.9.0
django-cors-headers>=4.3.1,<4.4.0
Pillow>=10.1.0,<10.2.0
dj-database-url>=2.1.0,<2.2.0
djangorestframework-simplejwt>=5.3.0,<5.4.0
django-environ==0.11.2
environ==1.0
pycparser==2.21
